chore: restructure folders system to increase clearity
