+++ src/view/viewState.ts
@@ -1,6 +1,6 @@
 export type NodeId = string;
-import type { NodeViewState } from "./types";
-import type { ViewState } from "./types";
+import type { NodeViewState } from "../types";
+import type { ViewState } from "../types";
 
 export function getOrCreateNodeViewState(
   viewState: ViewState,
